Implement mapreduce.reduce.shuffle.symlink-mapoutputs option.
When running Hadoop over a networked file system, such as NFS or Lustre,
there is no need to send map output data from mappers to reducers.  The
data is already stored in a network location that is available.  So instead
create a symlink to the mapoutput data.

This patch is a proof of concept and only works with 1 reducer.

See RESULTS.txt file for performance analysis of patch.

Signed-off-by: Albert Chu <chu11@llnl.gov>