Fix filename of udaMOFSupplier log file.
When the udaMOFSupplier log is created, it uses getlogin() to
determine the user associated with the file.  On many clusters
there is no terminal associated with the running process, so
this always results in a file name with a null in it, such as:

hadoop-(null)-udaMOFSupplier-myhost.log

This fix uses getuid/getpwuid instead of getlogin so the user
associated with the running process is used in the filename.

Signed-off-by: Albert Chu <chu11@llnl.gov>